wip: refactor whole module in lib
