+++ legacy-lib/src/backend/sticky_cookie.rs
