+++ legacy-lib/src/handler/error.rs
