+++ legacy-lib/src/handler/utils_synth_response.rs
