+++ rpxy-bin/Cargo.toml
@@ -12,15 +12,15 @@ publish = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-default = ["http3-s2n", "cache"]
-http3-quinn = ["rpxy-lib/http3-quinn"]
-http3-s2n = ["rpxy-lib/http3-s2n"]
-cache = ["rpxy-lib/cache"]
-native-roots = ["rpxy-lib/native-roots"]
+# default = ["http3-s2n", "cache"]
+# http3-quinn = ["rpxy-lib/http3-quinn"]
+# http3-s2n = ["rpxy-lib/http3-s2n"]
+# cache = ["rpxy-lib/cache"]
+# native-roots = ["rpxy-lib/native-roots"]
 
 [dependencies]
 rpxy-lib = { path = "../rpxy-lib/", default-features = false, features = [
-  "sticky-cookie",
+  # "sticky-cookie",
 ] }
 
 anyhow = "1.0.75"
